TextShow

Description

Efficiently convert from values to Text via Builders.

Since: 2

The TextShow classes

TextShow

class TextShow a where

Conversion of values to Text. Because there are both strict and lazy Text variants, the TextShow class deliberately avoids using Text in its functions. Instead, showbPrec, showb, and showbList all return Builder, an efficient intermediate form that can be converted to either kind of Text.

Builder is a Monoid, so it is useful to use the mappend (or <>) function to combine Builders when creating TextShow instances. As an example:

import Data.Monoid
import TextShow

data Example = Example Int Int
instance TextShow Example where
    showb (Example i1 i2) = showb i1 <> showbSpace <> showb i2

If you do not want to create TextShow instances manually, you can alternatively use the TextShow.TH module to automatically generate default TextShow instances using Template Haskell, or the TextShow.Generic module to quickly define TextShow instances using genericShowbPrec.

Since: 2

Minimal complete definition

showbPrec | showb

Methods

showbPrec :: Int -> a -> Builder

Convert a value to a Builder with the given predence.

Since: 2

showb :: a -> Builder

Converts a value to a strict Text. If you hand-define this, it should satisfy:

showb = showbPrec 0

Since: 2

showbList :: [a] -> Builder

Converts a list of values to a Builder. By default, this is defined as 'showbList = showbListWith showb, but it can be overridden to allow for specialized displaying of lists (e.g., lists of Chars).

Since: 2

showtPrec :: Int -> a -> Text

Converts a value to a strict Text with the given precedence. This can be overridden for efficiency, but it should satisfy:

showtPrec p = toStrict . showtlPrec p

Since: 3

showt :: a -> Text

Converts a value to a strict Text. This can be overridden for efficiency, but it should satisfy:

showt = showtPrec 0
showt = toStrict . showtl

The first equation is the default definition of showt.

Since: 3

showtList :: [a] -> Text

Converts a list of values to a strict Text. This can be overridden for efficiency, but it should satisfy:

showtList = toStrict . showtlList

Since: 3

showtlPrec :: Int -> a -> Text

Converts a value to a lazy Text with the given precedence. This can be overridden for efficiency, but it should satisfy:

showtlPrec p = toLazyText . showbPrec p

Since: 3

showtl :: a -> Text

Converts a value to a lazy Text. This can be overridden for efficiency, but it should satisfy:

showtl = showtlPrec 0
showtl = toLazyText . showb

The first equation is the default definition of showtl.

Since: 3

showtlList :: [a] -> Text

Converts a list of values to a lazy Text. This can be overridden for efficiency, but it should satisfy:

showtlList = toLazyText . showbList

Since: 3

showbParen :: Bool -> Builder -> Builder

Surrounds Builder output with parentheses if the Bool parameter is True.

Since: 2

showtParen :: Bool -> Text -> Text

Surrounds strict Text output with parentheses if the Bool parameter is True.

Since: 3.4

showtlParen :: Bool -> Text -> Text

Surrounds lazy Text output with parentheses if the Bool parameter is True.

Since: 3.4

showbCommaSpace :: Builder

Construct a Builder containing a comma followed by a space.

Since: 3.6

showtCommaSpace :: Text

Construct a strict Text containing a comma followed by a space.

Since: 3.6

showtlCommaSpace :: Text

Construct a lazy Text containing a comma followed by a space.

Since: 3.6

showbSpace :: Builder

Construct a Builder containing a single space character.

Since: 2

showtSpace :: Text

Construct a strict Text containing a single space character.

Since: 3.4

showtlSpace :: Text

Construct a lazy Text containing a single space character.

Since: 3.4

TextShow1

class TextShow1 f where

Lifting of the TextShow class to unary type constructors.

Since: 2

Minimal complete definition

liftShowbPrec

Methods

liftShowbPrec :: (Int -> a -> Builder) -> ([a] -> Builder) -> Int -> f a -> Builder

showbPrec function for an application of the type constructor based on showbPrec and showbList functions for the argument type.

Since: 3

liftShowbList :: (Int -> a -> Builder) -> ([a] -> Builder) -> [f a] -> Builder

showbList function for an application of the type constructor based on showbPrec and showbList functions for the argument type. The default implementation using standard list syntax is correct for most types.

Since: 3

showbPrec1 :: (TextShow1 f, TextShow a) => Int -> f a -> Builder

Lift the standard showbPrec and showbList functions through the type constructor.

Since: 2

showbUnaryWith :: (Int -> a -> Builder) -> Builder -> Int -> a -> Builder

showbUnaryWith sp n p x produces the Builder representation of a unary data constructor with name n and argument x, in precedence context p, using the function sp to show occurrences of the type argument.

Since: 2

liftShowtPrec :: TextShow1 f => (Int -> a -> Text) -> ([a] -> Text) -> Int -> f a -> Text

showtPrec function for an application of the type constructor based on showtPrec and showtList functions for the argument type.

The current implementation is based on liftShowbPrec internally.

Since: 3.4

liftShowtlPrec :: TextShow1 f => (Int -> a -> Text) -> ([a] -> Text) -> Int -> f a -> Text

showtlPrec function for an application of the type constructor based on showtlPrec and showtlList functions for the argument type.

The current implementation is based on liftShowbPrec internally.

Since: 3.4

TextShow2

class TextShow2 f where

Lifting of the TextShow class to binary type constructors.

Since: 2

Minimal complete definition

liftShowbPrec2

Methods

liftShowbPrec2 :: (Int -> a -> Builder) -> ([a] -> Builder) -> (Int -> b -> Builder) -> ([b] -> Builder) -> Int -> f a b -> Builder

showbPrec function for an application of the type constructor based on showbPrec and showbList functions for the argument types.

Since: 3

liftShowbList2 :: (Int -> a -> Builder) -> ([a] -> Builder) -> (Int -> b -> Builder) -> ([b] -> Builder) -> [f a b] -> Builder

showbList function for an application of the type constructor based on showbPrec and showbList functions for the argument types. The default implementation using standard list syntax is correct for most types.

Since: 3

showbPrec2 :: (TextShow2 f, TextShow a, TextShow b) => Int -> f a b -> Builder

Lift two showbPrec functions through the type constructor.

Since: 2

showbBinaryWith :: (Int -> a -> Builder) -> (Int -> b -> Builder) -> Builder -> Int -> a -> b -> Builder

showbBinaryWith sp n p x y produces the Builder representation of a binary data constructor with name n and arguments x and y, in precedence context p, using the functions sp1 and sp2 to show occurrences of the type arguments.

Since: 2

liftShowtPrec2 :: TextShow2 f => (Int -> a -> Text) -> ([a] -> Text) -> (Int -> b -> Text) -> ([b] -> Text) -> Int -> f a b -> Text

showtPrec function for an application of the type constructor based on showtPrec and showtList functions for the argument type.

The current implementation is based on liftShowbPrec2 internally.

Since: 3.4

liftShowtlPrec2 :: TextShow2 f => (Int -> a -> Text) -> ([a] -> Text) -> (Int -> b -> Text) -> ([b] -> Text) -> Int -> f a b -> Text

showtlPrec function for an application of the type constructor based on showtlPrec and showtlList functions for the argument type.

The current implementation is based on liftShowbPrec2 internally.

Since: 3.4

Builders

The Builder type

data Builder :: *

A Builder is an efficient way to build lazy Text values. There are several functions for constructing builders, but only one to inspect them: to extract any data, you have to turn them into lazy Text values using toLazyText.

Internally, a builder constructs a lazy Text by filling arrays piece by piece. As each buffer is filled, it is 'popped' off, to become a new chunk of the resulting lazy Text. All this is hidden from the user of the Builder.

toText :: Builder -> Text

Convert a Builder to a strict Text.

Since: 2

toLazyText :: Builder -> Text

O(n). Extract a lazy Text from a Builder with a default buffer size. The construction work takes place if and when the relevant part of the lazy Text is demanded.

toLazyTextWith :: Int -> Builder -> Text

O(n). Extract a lazy Text from a Builder, using the given size for the initial buffer. The construction work takes place if and when the relevant part of the lazy Text is demanded.

If the initial buffer is too small to hold all data, subsequent buffers will be the default buffer size.

toString :: Builder -> String

Convert a Builder to a String (without surrounding it with double quotes, as show would).

Since: 2

Constructing Builders

singleton :: Char -> Builder

O(1). A Builder taking a single character, satisfying

toLazyText (singleton c) = singleton c

fromText :: Text -> Builder

O(1). A Builder taking a Text, satisfying

toLazyText (fromText t) = fromChunks [t]

fromLazyText :: Text -> Builder

O(1). A Builder taking a lazy Text, satisfying

toLazyText (fromLazyText t) = t

fromString :: String -> Builder

O(1). A Builder taking a String, satisfying

toLazyText (fromString s) = fromChunks [S.pack s]

Flushing the buffer state

flush :: Builder

O(1). Pop the strict Text we have constructed so far, if any, yielding a new chunk in the result lazy Text.

Builder utility functions

lengthB :: Builder -> Int64

Computes the length of a Builder.

Since: 2

unlinesB :: [Builder] -> Builder

Merges several Builders, separating them by newlines.

Since: 2

unwordsB :: [Builder] -> Builder

Merges several Builders, separating them by spaces.

Since: 2

Printing values

printT :: TextShow a => a -> IO ()

Writes a value's strict Text representation to the standard output, followed by a newline.

Since: 2

printTL :: TextShow a => a -> IO ()

Writes a value's lazy Text representation to the standard output, followed by a newline.

Since: 2

hPrintT :: TextShow a => Handle -> a -> IO ()

Writes a value's strict Text representation to a file handle, followed by a newline.

Since: 2

hPrintTL :: TextShow a => Handle -> a -> IO ()

Writes a value's lazy Text representation to a file handle, followed by a newline.

Since: 2

Conversions

Conversion between TextShow and string Show

newtype FromStringShow a

The TextShow instance for FromStringShow is based on its String Show instance. That is,

showbPrec p (FromStringShow x) = showsToShowb showsPrec p x

Since: 2

Constructors

FromStringShow

Fields

fromStringShow :: a

newtype FromTextShow a

The String Show instance for FromTextShow is based on its TextShow instance. That is,

showsPrec p (FromTextShow x) = showbToShows showbPrec p x

Since: 2

Constructors

FromTextShow

Fields

fromTextShow :: a

newtype FromStringShow1 f a

The TextShow1 instance for FromStringShow1 is based on its String Show1 instance. That is,

liftShowbPrec sp sl p (FromStringShow1 x) =
    showsPrecToShowbPrec (liftShowsPrec (showbPrecToShowsPrec sp)
                                            (showbToShows         sl))
                           p x

Since: 3

Constructors

FromStringShow1

Fields

fromStringShow1 :: f a

newtype FromTextShow1 f a

The String Show1 instance for FromTextShow1 is based on its TextShow1 instance. That is,

liftShowsPrec sp sl p (FromTextShow1 x) =
    showbPrecToShowsPrec (liftShowbPrec (showsPrecToShowbPrec sp)
                                            (showsToShowb         sl))
                           p x

Since: 3

Constructors

FromTextShow1

Fields

fromTextShow1 :: f a

newtype FromStringShow2 f a b

The TextShow2 instance for FromStringShow2 is based on its String Show2 instance. That is,

liftShowbPrec2 sp1 sl1 sp2 sl2 p (FromStringShow2 x) =
    showsPrecToShowbPrec (liftShowsPrec2 (showbPrecToShowsPrec sp1)
                                             (showbToShows         sl1)
                                             (showbPrecToShowsPrec sp2)
                                             (showbToShows         sl2))
                           p x

Since: 3

Constructors

FromStringShow2

Fields

fromStringShow2 :: f a b

newtype FromTextShow2 f a b

The String Show2 instance for FromTextShow2 is based on its TextShow2 instance. That is,

liftShowsPrec2 sp1 sl1 sp2 sl2 p (FromTextShow2 x) =
    showbPrecToShowsPrec (liftShowbPrec2 (showsPrecToShowbPrec sp1)
                                             (showsToShowb         sl1)
                                             (showsPrecToShowbPrec sp2)
                                             (showsToShowb         sl2))
                           p x

Since: 3

Constructors

FromTextShow2

Fields

fromTextShow2 :: f a b

showsPrecToShowbPrec :: (Int -> a -> ShowS) -> Int -> a -> Builder

Convert a precedence-aware ShowS-based show function to a Builder-based one.

Since: 3

showsToShowb :: (a -> ShowS) -> a -> Builder

Convert a ShowS-based show function to a Builder-based one.

Since: 3

showbPrecToShowsPrec :: (Int -> a -> Builder) -> Int -> a -> ShowS

Convert a precedence-aware Builder-based show function to a ShowS-based one.

Since: 3

showbToShows :: (a -> Builder) -> a -> ShowS

Convert a Builder-based show function to a ShowS-based one.

Since: 3

Conversions between Builder, strict Text, and lazy Text

showtPrecToShowbPrec :: (Int -> a -> Text) -> Int -> a -> Builder

Convert a precedence-aware, strict Text-based show function to a Builder-based one.

Since: 3.4

showtlPrecToShowbPrec :: (Int -> a -> Text) -> Int -> a -> Builder

Convert a precedence-aware, lazy Text-based show function to a Builder-based one.

Since: 3.4

showtToShowb :: (a -> Text) -> a -> Builder

Convert a strict Text-based show function to a Builder-based one.

Since: 3.4

showtlToShowb :: (a -> Text) -> a -> Builder

Convert a lazy Text-based show function to a Builder-based one.

Since: 3.4

showbPrecToShowtPrec :: (Int -> a -> Builder) -> Int -> a -> Text

Convert a precedence-aware Builder-based show function to a strict Text-based one.

Since: 3.4

showbPrecToShowtlPrec :: (Int -> a -> Builder) -> Int -> a -> Text

Convert a precedence-aware Builder-based show function to a lazy Text-based one.

Since: 3.4

showbToShowt :: (a -> Builder) -> a -> Text

Convert a Builder-based show function to a strict Text-based one.

Since: 3

showbToShowtl :: (a -> Builder) -> a -> Text

Convert a Builder-based show function to a lazy Text-based one.

Since: 3