TextShow.Data.Array

Description

Provides TextShow instances for Array types, as well as the showbIArrayPrec function.

Since: 2

Documentation

showbIArrayPrec :: (IArray a e, Ix i, TextShow i, TextShow e) => Int -> a i e -> Builder

Convert an IArray instance to a Builder with the given precedence.

Since: 2

Orphan instances

(IArray UArray e, Ix i, TextShow i, TextShow e) => TextShow (UArray i e)
Since: 2

Methods

showbPrec :: Int -> UArray i e -> Builder

showb :: UArray i e -> Builder

showbList :: [UArray i e] -> Builder

showtPrec :: Int -> UArray i e -> Text

showt :: UArray i e -> Text

showtList :: [UArray i e] -> Text

showtlPrec :: Int -> UArray i e -> Text

showtl :: UArray i e -> Text

showtlList :: [UArray i e] -> Text

(TextShow i, TextShow e, Ix i) => TextShow (Array i e)
Since: 2

Methods

showbPrec :: Int -> Array i e -> Builder

showb :: Array i e -> Builder

showbList :: [Array i e] -> Builder

showtPrec :: Int -> Array i e -> Text

showt :: Array i e -> Text

showtList :: [Array i e] -> Text

showtlPrec :: Int -> Array i e -> Text

showtl :: Array i e -> Text

showtlList :: [Array i e] -> Text