TextShow.Data.Char

Description

TextShow instances and monomorphic functions for Char and String.

Since: 2

Documentation

showbChar :: Char -> Builder

Convert a Char to a Builder (surrounded by single quotes).

Since: 2

showbLitChar :: Char -> Builder

Convert a Char to a Builder (without single quotes).

Since: 2

showbString :: String -> Builder

Convert a String to a Builder (surrounded by double quotes).

Since: 2

showbLitString :: String -> Builder

Convert a String to a Builder (without double quotes).

Since: 2

showbGeneralCategory :: GeneralCategory -> Builder

Convert a GeneralCategory to a Builder.

Since: 2

asciiTabB :: Array Int Builder

A table of ASCII control characters that needs to be escaped with a backslash.

Since: 2

Orphan instances

TextShow Char
Since: 2

Methods

showbPrec :: Int -> Char -> Builder

showb :: Char -> Builder

showbList :: [Char] -> Builder

showtPrec :: Int -> Char -> Text

showt :: Char -> Text

showtList :: [Char] -> Text

showtlPrec :: Int -> Char -> Text

showtl :: Char -> Text

showtlList :: [Char] -> Text

TextShow GeneralCategory
Since: 2

Methods

showbPrec :: Int -> GeneralCategory -> Builder

showb :: GeneralCategory -> Builder

showbList :: [GeneralCategory] -> Builder

showtPrec :: Int -> GeneralCategory -> Text

showt :: GeneralCategory -> Text

showtList :: [GeneralCategory] -> Text

showtlPrec :: Int -> GeneralCategory -> Text

showtl :: GeneralCategory -> Text

showtlList :: [GeneralCategory] -> Text