TextShow.Data.Floating

Description

TextShow instances and monomorphic functions for floating-point types.

Since: 2

Documentation

showbRealFloatPrec :: RealFloat a => Int -> a -> Builder

Convert a RealFloat value to a Builder with the given precedence.

Since: 2

showbEFloat :: RealFloat a => Maybe Int -> a -> Builder

Show a signed RealFloat value using scientific (exponential) notation (e.g. 2.45e2, 1.5e-3).

In the call showbEFloat digs val, if digs is Nothing, the value is shown to full precision; if digs is Just d, then at most d digits after the decimal point are shown.

Since: 2

showbFFloat :: RealFloat a => Maybe Int -> a -> Builder

Show a signed RealFloat value using standard decimal notation (e.g. 245000, 0.0015).

In the call showbFFloat digs val, if digs is Nothing, the value is shown to full precision; if digs is Just d, then at most d digits after the decimal point are shown.

Since: 2

showbGFloat :: RealFloat a => Maybe Int -> a -> Builder

Show a signed RealFloat value using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

In the call showbGFloat digs val, if digs is Nothing, the value is shown to full precision; if digs is Just d, then at most d digits after the decimal point are shown.

Since: 2

showbFFloatAlt :: RealFloat a => Maybe Int -> a -> Builder

Show a signed RealFloat value using standard decimal notation (e.g. 245000, 0.0015).

This behaves as showFFloat, except that a decimal point is always guaranteed, even if not needed.

Since: 2

showbGFloatAlt :: RealFloat a => Maybe Int -> a -> Builder

Show a signed RealFloat value using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

This behaves as showFFloat, except that a decimal point is always guaranteed, even if not needed.

Since: 2

showbFPFormat :: FPFormat -> Builder

Convert an FPFormat value to a Builder.

Since: 2

data FPFormat :: *

Control the rendering of floating point numbers.

Constructors

Exponent
Scientific notation (e.g. 2.3e123).

Fixed
Standard decimal notation.

Generic
Use decimal notation for values between 0.1 and 9,999,999, and scientific notation otherwise.

formatRealFloatB

Arguments

:: RealFloat a

=> FPFormat
What notation to use.

-> Maybe Int
Number of decimal places to render.

-> a

-> Builder

Like formatRealFloatAltB, except that the decimal is only shown for arguments whose absolute value is between 0.1 and 9,999,999.

Since: 2

formatRealFloatAltB

Arguments

:: RealFloat a

=> FPFormat
What notation to use.

-> Maybe Int
Number of decimal places to render.

-> Bool
Should a decimal point always be shown?

-> a

-> Builder

Converts a RealFloat value to a Builder, specifying if a decimal point should always be shown.

Since: 2

Orphan instances

TextShow Double
Since: 2

Methods

showbPrec :: Int -> Double -> Builder

showb :: Double -> Builder

showbList :: [Double] -> Builder

showtPrec :: Int -> Double -> Text

showt :: Double -> Text

showtList :: [Double] -> Text

showtlPrec :: Int -> Double -> Text

showtl :: Double -> Text

showtlList :: [Double] -> Text

TextShow Float
Since: 2

Methods

showbPrec :: Int -> Float -> Builder

showb :: Float -> Builder

showbList :: [Float] -> Builder

showtPrec :: Int -> Float -> Text

showt :: Float -> Text

showtList :: [Float] -> Text

showtlPrec :: Int -> Float -> Text

showtl :: Float -> Text

showtlList :: [Float] -> Text

TextShow FPFormat
Since: 2

Methods

showbPrec :: Int -> FPFormat -> Builder

showb :: FPFormat -> Builder

showbList :: [FPFormat] -> Builder

showtPrec :: Int -> FPFormat -> Text

showt :: FPFormat -> Text

showtList :: [FPFormat] -> Text

showtlPrec :: Int -> FPFormat -> Text

showtl :: FPFormat -> Text

showtlList :: [FPFormat] -> Text