TextShow.Data.Integral

Description

TextShow instances and monomorphic functions for integral types.

Since: 2

Documentation

showbIntegralPrec :: Integral a => Int -> a -> Builder

Convert an Integral type to a Builder with the given precedence.

Since: 2

showbIntAtBase :: (Integral a, TextShow a) => a -> (Int -> Char) -> a -> Builder

Shows a non-negative Integral number using the base specified by the first argument, and the character representation specified by the second.

Since: 2

showbBin :: (Integral a, TextShow a) => a -> Builder

Show non-negative Integral numbers in base 2.

Since: 2

showbHex :: (Integral a, TextShow a) => a -> Builder

Show non-negative Integral numbers in base 16.

Since: 2

showbOct :: (Integral a, TextShow a) => a -> Builder

Show non-negative Integral numbers in base 8.

Since: 2

Orphan instances

TextShow Int
Since: 2

Methods

showbPrec :: Int -> Int -> Builder

showb :: Int -> Builder

showbList :: [Int] -> Builder

showtPrec :: Int -> Int -> Text

showt :: Int -> Text

showtList :: [Int] -> Text

showtlPrec :: Int -> Int -> Text

showtl :: Int -> Text

showtlList :: [Int] -> Text

TextShow Int8
Since: 2

Methods

showbPrec :: Int -> Int8 -> Builder

showb :: Int8 -> Builder

showbList :: [Int8] -> Builder

showtPrec :: Int -> Int8 -> Text

showt :: Int8 -> Text

showtList :: [Int8] -> Text

showtlPrec :: Int -> Int8 -> Text

showtl :: Int8 -> Text

showtlList :: [Int8] -> Text

TextShow Int16
Since: 2

Methods

showbPrec :: Int -> Int16 -> Builder

showb :: Int16 -> Builder

showbList :: [Int16] -> Builder

showtPrec :: Int -> Int16 -> Text

showt :: Int16 -> Text

showtList :: [Int16] -> Text

showtlPrec :: Int -> Int16 -> Text

showtl :: Int16 -> Text

showtlList :: [Int16] -> Text

TextShow Int32
Since: 2

Methods

showbPrec :: Int -> Int32 -> Builder

showb :: Int32 -> Builder

showbList :: [Int32] -> Builder

showtPrec :: Int -> Int32 -> Text

showt :: Int32 -> Text

showtList :: [Int32] -> Text

showtlPrec :: Int -> Int32 -> Text

showtl :: Int32 -> Text

showtlList :: [Int32] -> Text

TextShow Int64
Since: 2

Methods

showbPrec :: Int -> Int64 -> Builder

showb :: Int64 -> Builder

showbList :: [Int64] -> Builder

showtPrec :: Int -> Int64 -> Text

showt :: Int64 -> Text

showtList :: [Int64] -> Text

showtlPrec :: Int -> Int64 -> Text

showtl :: Int64 -> Text

showtlList :: [Int64] -> Text

TextShow Integer
Since: 2

Methods

showbPrec :: Int -> Integer -> Builder

showb :: Integer -> Builder

showbList :: [Integer] -> Builder

showtPrec :: Int -> Integer -> Text

showt :: Integer -> Text

showtList :: [Integer] -> Text

showtlPrec :: Int -> Integer -> Text

showtl :: Integer -> Text

showtlList :: [Integer] -> Text

TextShow Word
Since: 2

Methods

showbPrec :: Int -> Word -> Builder

showb :: Word -> Builder

showbList :: [Word] -> Builder

showtPrec :: Int -> Word -> Text

showt :: Word -> Text

showtList :: [Word] -> Text

showtlPrec :: Int -> Word -> Text

showtl :: Word -> Text

showtlList :: [Word] -> Text

TextShow Word8
Since: 2

Methods

showbPrec :: Int -> Word8 -> Builder

showb :: Word8 -> Builder

showbList :: [Word8] -> Builder

showtPrec :: Int -> Word8 -> Text

showt :: Word8 -> Text

showtList :: [Word8] -> Text

showtlPrec :: Int -> Word8 -> Text

showtl :: Word8 -> Text

showtlList :: [Word8] -> Text

TextShow Word16
Since: 2

Methods

showbPrec :: Int -> Word16 -> Builder

showb :: Word16 -> Builder

showbList :: [Word16] -> Builder

showtPrec :: Int -> Word16 -> Text

showt :: Word16 -> Text

showtList :: [Word16] -> Text

showtlPrec :: Int -> Word16 -> Text

showtl :: Word16 -> Text

showtlList :: [Word16] -> Text

TextShow Word32
Since: 2

Methods

showbPrec :: Int -> Word32 -> Builder

showb :: Word32 -> Builder

showbList :: [Word32] -> Builder

showtPrec :: Int -> Word32 -> Text

showt :: Word32 -> Text

showtList :: [Word32] -> Text

showtlPrec :: Int -> Word32 -> Text

showtl :: Word32 -> Text

showtlList :: [Word32] -> Text

TextShow Word64
Since: 2

Methods

showbPrec :: Int -> Word64 -> Builder

showb :: Word64 -> Builder

showbList :: [Word64] -> Builder

showtPrec :: Int -> Word64 -> Text

showt :: Word64 -> Text

showtList :: [Word64] -> Text

showtlPrec :: Int -> Word64 -> Text

showtl :: Word64 -> Text

showtlList :: [Word64] -> Text