TextShow.Debug.Trace.TH

Description

Functions that splice traces into source code which take an arbitrary data type or data family instance as an argument (even if it is not an instance of TextShow). You need to enable the TemplateHaskell language extension in order to use this module.

Since: 2

Documentation

makeTraceTextShow :: Name -> Q Exp

Generates a lambda expression which behaves like traceTextShow (without requiring a TextShow instance).

Since: 2

makeTraceTextShowId :: Name -> Q Exp

Generates a lambda expression which behaves like traceTextShowId (without requiring a TextShow instance).

Since: 2

makeTraceTextShowM :: Name -> Q Exp

Generates a lambda expression which behaves like traceTextShowM (without requiring a TextShow instance).

Since: 2