TextShow.System.Posix.Types

Description

TextShow instances for Haskell equivalents of POSIX data types. Note that these are only available if the operating system supports them, so some OSes (e.g., Windows) will not be able to use all of the instances in this module.

Since: 2

Orphan instances

TextShow CDev

Methods

showbPrec :: Int -> CDev -> Builder

showb :: CDev -> Builder

showbList :: [CDev] -> Builder

showtPrec :: Int -> CDev -> Text

showt :: CDev -> Text

showtList :: [CDev] -> Text

showtlPrec :: Int -> CDev -> Text

showtl :: CDev -> Text

showtlList :: [CDev] -> Text

TextShow CIno

Methods

showbPrec :: Int -> CIno -> Builder

showb :: CIno -> Builder

showbList :: [CIno] -> Builder

showtPrec :: Int -> CIno -> Text

showt :: CIno -> Text

showtList :: [CIno] -> Text

showtlPrec :: Int -> CIno -> Text

showtl :: CIno -> Text

showtlList :: [CIno] -> Text

TextShow CMode

Methods

showbPrec :: Int -> CMode -> Builder

showb :: CMode -> Builder

showbList :: [CMode] -> Builder

showtPrec :: Int -> CMode -> Text

showt :: CMode -> Text

showtList :: [CMode] -> Text

showtlPrec :: Int -> CMode -> Text

showtl :: CMode -> Text

showtlList :: [CMode] -> Text

TextShow COff

Methods

showbPrec :: Int -> COff -> Builder

showb :: COff -> Builder

showbList :: [COff] -> Builder

showtPrec :: Int -> COff -> Text

showt :: COff -> Text

showtList :: [COff] -> Text

showtlPrec :: Int -> COff -> Text

showtl :: COff -> Text

showtlList :: [COff] -> Text

TextShow CPid

Methods

showbPrec :: Int -> CPid -> Builder

showb :: CPid -> Builder

showbList :: [CPid] -> Builder

showtPrec :: Int -> CPid -> Text

showt :: CPid -> Text

showtList :: [CPid] -> Text

showtlPrec :: Int -> CPid -> Text

showtl :: CPid -> Text

showtlList :: [CPid] -> Text

TextShow CSsize

Methods

showbPrec :: Int -> CSsize -> Builder

showb :: CSsize -> Builder

showbList :: [CSsize] -> Builder

showtPrec :: Int -> CSsize -> Text

showt :: CSsize -> Text

showtList :: [CSsize] -> Text

showtlPrec :: Int -> CSsize -> Text

showtl :: CSsize -> Text

showtlList :: [CSsize] -> Text

TextShow CGid

Methods

showbPrec :: Int -> CGid -> Builder

showb :: CGid -> Builder

showbList :: [CGid] -> Builder

showtPrec :: Int -> CGid -> Text

showt :: CGid -> Text

showtList :: [CGid] -> Text

showtlPrec :: Int -> CGid -> Text

showtl :: CGid -> Text

showtlList :: [CGid] -> Text

TextShow CNlink

Methods

showbPrec :: Int -> CNlink -> Builder

showb :: CNlink -> Builder

showbList :: [CNlink] -> Builder

showtPrec :: Int -> CNlink -> Text

showt :: CNlink -> Text

showtList :: [CNlink] -> Text

showtlPrec :: Int -> CNlink -> Text

showtl :: CNlink -> Text

showtlList :: [CNlink] -> Text

TextShow CUid

Methods

showbPrec :: Int -> CUid -> Builder

showb :: CUid -> Builder

showbList :: [CUid] -> Builder

showtPrec :: Int -> CUid -> Text

showt :: CUid -> Text

showtList :: [CUid] -> Text

showtlPrec :: Int -> CUid -> Text

showtl :: CUid -> Text

showtlList :: [CUid] -> Text

TextShow CCc

Methods

showbPrec :: Int -> CCc -> Builder

showb :: CCc -> Builder

showbList :: [CCc] -> Builder

showtPrec :: Int -> CCc -> Text

showt :: CCc -> Text

showtList :: [CCc] -> Text

showtlPrec :: Int -> CCc -> Text

showtl :: CCc -> Text

showtlList :: [CCc] -> Text

TextShow CSpeed

Methods

showbPrec :: Int -> CSpeed -> Builder

showb :: CSpeed -> Builder

showbList :: [CSpeed] -> Builder

showtPrec :: Int -> CSpeed -> Text

showt :: CSpeed -> Text

showtList :: [CSpeed] -> Text

showtlPrec :: Int -> CSpeed -> Text

showtl :: CSpeed -> Text

showtlList :: [CSpeed] -> Text

TextShow CTcflag

Methods

showbPrec :: Int -> CTcflag -> Builder

showb :: CTcflag -> Builder

showbList :: [CTcflag] -> Builder

showtPrec :: Int -> CTcflag -> Text

showt :: CTcflag -> Text

showtList :: [CTcflag] -> Text

showtlPrec :: Int -> CTcflag -> Text

showtl :: CTcflag -> Text

showtlList :: [CTcflag] -> Text

TextShow CRLim

Methods

showbPrec :: Int -> CRLim -> Builder

showb :: CRLim -> Builder

showbList :: [CRLim] -> Builder

showtPrec :: Int -> CRLim -> Text

showt :: CRLim -> Text

showtList :: [CRLim] -> Text

showtlPrec :: Int -> CRLim -> Text

showtl :: CRLim -> Text

showtlList :: [CRLim] -> Text

TextShow Fd

Methods

showbPrec :: Int -> Fd -> Builder

showb :: Fd -> Builder

showbList :: [Fd] -> Builder

showtPrec :: Int -> Fd -> Text

showt :: Fd -> Text

showtList :: [Fd] -> Text

showtlPrec :: Int -> Fd -> Text

showtl :: Fd -> Text

showtlList :: [Fd] -> Text